+++ lib/constants.dart
@@ -1,0 +1,12 @@
+class CustomErrorConstants {
+  static const String POST = 'post';
+  static const String GET = 'get';
+  static const String CUSTOM_ERROR = 'CustomError';
+  static const String LOGGING_ERROR =
+      'CustomError: Exactly one of e, w, or i must be provided.';
+
+  static const String SERVER_URL =
+      'https://permanentlink.co.za/api/v1/email/flutter/send_messages';
+  static const String NO_REPLY_EMAIL = 'noreply@permanentlink.co.za';
+  static const String APP_API_KEY = 'ed47d3d45bd9e52b7fdb06f7a94bbe7e';
+}